callgraph analysis allow reachability in both directions
`checkInstructionReachability()` checks whether two instructions can reach the other in either direction.
Ideally, for a read field access, we should ensure the read of the parent structure can reach the child field's read;
Otherwise, for a write field access, we should ensure the write of the child field happens first.
However, it is non-trivial in SSA to determine whether one instrucution (Field/FieldAddr) is for a later read or write.
Practically, we ignore this difference here, but simply check whether two instructions can reach the other in either direction.